requirements.txt
Flask==2.2.5
mysql-connector-python==8.3.0
deepface==0.0.79
Pillow==10.2.0
requests==2.31.0
boto3
botocore